+++ tests/unit/views/TreeView.spec.ts
@@ -1,9 +1,10 @@
-import { mount, createLocalVue, Wrapper } from '@vue/test-utils'
+import { mount, createLocalVue, Wrapper, shallowMount } from '@vue/test-utils'
 import TreeView from '@/views/TreeView.vue'
 import Vue from 'vue'
 import Vuex from 'vuex'
 import ApplicationState from '@/types/ApplicationState'
 import emptyState from '@/store/state'
+import SearchComponent from '@/components/search/SearchComponent.vue'
 
 const localVue = createLocalVue()
 localVue.use(Vuex)
@@ -15,10 +16,19 @@ describe('TreeView.vue', () => {
   let store: any
   let treeUpdate = jest.fn()
   let treeOpenUpdate = jest.fn()
+
+  const updateSearchTerm = jest.fn()
+  const filterSections = jest.fn()
+  const filterSubsections = jest.fn()
+  const loadGridVariables = jest.fn()
+
   let actions = {
     loadSections: jest.fn(),
     loadSubSections: jest.fn(),
-    loadSectionTree: jest.fn()
+    loadSectionTree: jest.fn(),
+    filterSections,
+    filterSubsections,
+    loadGridVariables
   }
 
   beforeEach(() => {
@@ -43,7 +53,8 @@ describe('TreeView.vue', () => {
       actions,
       mutations: {
         updateTreeSelection: treeUpdate,
-        updateTreeOpenSection: treeOpenUpdate
+        updateTreeOpenSection: treeOpenUpdate,
+        updateSearchTerm
       }
     })
 
@@ -69,4 +80,15 @@ describe('TreeView.vue', () => {
     wrapper.find('[title="child"]').trigger('click')
     expect(treeUpdate.mock.calls).toEqual([[{ 'treeOpenSection': -1, 'treeSelected': -1 }, 10]])
   })
+  it('updates search term and dispatches filter actions when the search term changes', () => {
+    const wrapper = shallowMount(TreeView, { store, localVue })
+
+    wrapper.find(SearchComponent).vm.$emit('searchChanged', 'mini')
+
+    expect(updateSearchTerm).toHaveBeenCalled()
+    expect(filterSections).toHaveBeenCalled()
+    expect(filterSubsections).toHaveBeenCalled()
+    expect(loadGridVariables).toHaveBeenCalled()
+  })
+
 })
